Create utilities file
